Divide SwaggerTest on to three dependences tests. Plus a little refactoring.
