+++ src/test/java/ui/model/api/SwaggerSettingPage.java
@@ -1,4 +1,4 @@
-package ui.model;
+package ui.model.api;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
