+++ src/test/java/ui/model/base/SidePanel.java
@@ -5,9 +5,9 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
-import ui.model.AddPluginsPage;
-import ui.model.PermalinkSettingsPage;
-import ui.model.SwaggerSettingPage;
+import ui.model.plugins.AddPluginsPage;
+import ui.model.settings.PermalinkSettingsPage;
+import ui.model.api.SwaggerSettingPage;
 import ui.model.appearance.AppearancePage;
 import ui.model.comments.CommentsPage;
 import ui.model.pages.NewPagePage;
